use new gecko codes component
